Add files via upload
Update: Solved the questions and put documentation.

Answer for question before: Because BTB is empty initially, it will store the first BTA it meets into it. The second time the same branch happens, the update of valid bit (which would cause a MISS) and HIT happens in the same time. (If I am correct)